terraform/layers/backend/src/lambda_functions/s3_ingestion/requirements.txt
boto3
pdfplumber
python-docx
opensearch-py
requests
requests-aws4auth
